+++ shiver/out_flame.cpp
@@ -1,4 +1,0 @@
-#include <iostream>
-#include <cstdint>
-#include <array>
-#include <vector>