+++ src/table.cpp
@@ -1,5 +1,5 @@
-#include "include/table.h"
-#include "include/lexer.h"
+#include "../include/table.h"
+#include "../include/lexer.h"
 #include <ranges>
 #include <string>
 #include <utility>
@@ -57,7 +57,7 @@ token_type search_type_scope(const std::string &name, unsigned int lvl) {
 
 void insert(const std::string &name,token_type type, token_value val, bool is_const, u64 size, bool is_array, bool comptime, bool is_vector, bool isptr, const std::string &modname, bool ismov, bool isref) {
     if(table.empty()) [[unlikely]] return;
-    table[table.size()-1].insert_or_assign(name, symbol{type, std::move(val), is_const, size, is_array, comptime,name, is_vector, isptr, ismov, isref, modname});
+    table[table.size()-1].insert_or_assign(name, symbol{type, std::move(val), is_const, size, is_array, comptime,name, is_vector, isptr, modname, ismov});
 }
 
 void insert_top(const std::string &name,token_type type,token_value val, bool is_const, u64 size,bool is_array, bool comptime, bool is_vector, bool isptr, const std::string modname) {